Resolvendo issues do Sonar

+++ src/main/java/br/com/fiap/grupo1/microservicos/perfil/PerfilService.java
@@ -58,6 +58,6 @@ public class PerfilService {
 	}
 
 	private BigDecimal valorTotal(Stream<Pagamento> pagamentos) {
-		return (BigDecimal) pagamentos.map(Pagamento::getValor).collect(Pagamento.somaBigDecima());
+		return pagamentos.map(Pagamento::getValor).collect(Pagamento.somaBigDecima());
 	}
 }
